third_party/WebKit/LayoutTests/http/tests/security/vibration/vibrate-in-cross-origin-iframe-blocked-expected.txt

CONSOLE ERROR: line 10: Blocked call to navigator.vibrate inside a cross-origin iframe because the frame has never been activated by the user: https://www.chromestatus.com/feature/5682658461876224.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
This is a testharness.js-based test.
PASS A cross-origin iframe may not call navigator.vibrate: https://www.chromestatus.com/feature/5682658461876224.
Harness: the test ran to completion.
